requirement.txt
validators
youtube_transcript_api
langchain
langchain_core
langchain_community
langchain_groq
python-dotenv
streamlit
pytube
unstructured